customer side

+++ FOODIVE/admin/delete_category.aspx.cs
@@ -17,11 +17,16 @@ namespace FOODIVE.admin
         {
             if (Request.QueryString["cat_del"] != null)
             {
-                string query = "DELETE FROM category WHERE cate_id = " + Request.QueryString["cat_del"];
                 con.Open();
-                SqlCommand cmd = new SqlCommand(query, con);
-                if (cmd.ExecuteNonQuery() != 0)
-                    Response.Redirect("all_category.aspx");
+                string q1 = "delete from dishes where cate_id = " + Request.QueryString["cat_del"];
+                SqlCommand cmd1 = new SqlCommand(q1, con);
+                if (cmd1.ExecuteNonQuery() != 0)
+                {
+                    string query = "DELETE FROM category WHERE cate_id = " + Request.QueryString["cat_del"];
+                    SqlCommand cmd = new SqlCommand(query, con);
+                    if (cmd.ExecuteNonQuery() != 0)
+                        Response.Redirect("all_category.aspx");
+                }
                 con.Close();
             }
             else
